requirements.txt
Flask==3.0.2
python-dotenv==1.0.1 
anthropic==0.45.2